third_party/WebKit/LayoutTests/editing/deleting/delete-character-003-expected.txt

PASS sample.innerHTML is "a<br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 2
PASS successfullyParsed is true

TEST COMPLETE
